refactor: consolidate sorting and filtering utilities for DTSI client person data table, updating imports across country-specific implementations
