+++ src/components/app/pagePoliticians/ca/index.tsx
@@ -1,5 +1,5 @@
 import { CaDTSIClientPersonDataTable } from '@/components/app/dtsiClientPersonDataTable/ca'
-import { DTSIPersonDataTablePeople } from '@/components/app/dtsiClientPersonDataTable/common/sortPeople'
+import { DTSIPersonDataTablePeople } from '@/components/app/dtsiClientPersonDataTable/common/utils'
 import { PagePoliticiansLayout } from '@/components/app/pagePoliticians/common/layout'
 import { PageSubTitle } from '@/components/ui/pageSubTitle'
 import { PageTitle } from '@/components/ui/pageTitleText'
