+++ src/components/app/pagePoliticians/gb/index.tsx
@@ -1,4 +1,4 @@
-import { DTSIPersonDataTablePeople } from '@/components/app/dtsiClientPersonDataTable/common/sortPeople'
+import { DTSIPersonDataTablePeople } from '@/components/app/dtsiClientPersonDataTable/common/utils'
 import { GbDTSIClientPersonDataTable } from '@/components/app/dtsiClientPersonDataTable/gb'
 import { PagePoliticiansLayout } from '@/components/app/pagePoliticians/common/layout'
 import { PageSubTitle } from '@/components/ui/pageSubTitle'
